demo/requirements.txt
azure-ai-projects
azure-ai-agents
azure-ai-inference
azure-identity
opentelemetry-sdk
azure-core-tracing-opentelemetry
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp

logfire
openai
openai-agents